Add unlock section components and create barrel file
