+++ src/pages/SparkSwapWebsite/components/index.tsx
@@ -1,7 +1,8 @@
 export { default as Navigation } from './Navigation'
 export { default as Hero } from './Hero'
-export { default as Innovative } from './Innovative'
-export { default as KeyFeatures } from './KeyFeatures'
+export { default as Unlock } from './Unlock'
+export { default as Platform } from './Platform'
+// export { default as KeyFeatures } from './KeyFeatures'
 // export { default as SupportedCrypto } from './SupportedCrypto'
 export { default as PoweredByBSC } from './PoweredByBSC'
 export { default as SfuelToken } from './SfuelToken'
